# server/userapi/user_api_openapi.yaml
openapi: 3.0.3
info:
  title: Sora Mail User API
  description: |
    REST API for user mailbox access via HTTP, providing a modern alternative to IMAP/POP3 protocols.
    Features stateless JWT-based authentication for web and mobile clients.
  version: 1.0.0
  contact:
    name: Sora Mail Support
    url: https://github.com/migadu/sora
servers:
  - url: /user/v1
    description: User API version 1
...
paths:
...
  /messages/{id}/body:
    get:
      tags:
        - Messages
      summary: Get message body
      description: Retrieve message body in text or HTML format
      operationId: getMessageBody
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
        - name: format
          in: query
          description: Preferred format
          schema:
            type: string
            enum: [text, html]
            default: html
      responses:
        '200':
          description: Message body
          content:
            text/plain:
              schema:
                type: string
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /messages/{id}/raw:
    get:
      tags:
        - Messages
      summary: Get raw message
      description: Retrieve message in RFC822 format
      operationId: getMessageRaw
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Raw RFC822 message
          content:
            message/rfc822:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /user/v1/auth/login
  parameters:
...
    MessageId:
      name: id
      in: path
      required: true
      description: Message ID
      schema:
        type: integer
        format: int64
...
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found